moved pilldao sublcass in its own class
